project organized and tile choice optimized
