+++ src/java/SalamanderManiacs/blockminerClient/Graph/Image.java
@@ -1,0 +1,4 @@
+package SalamanderManiacs.blockminerClient.Graph;
+
+public class Image {
+}
